pyproject.toml
[project]
name = "fast-zero"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.14.1",
    "fastapi[standard]>=0.115.2",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.37",
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint.pycodestyle]
ignore-overlong-task-comments = true

[tool.ruff.lint]
preview = true
task-tags = ["HACK"]
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'fastapi dev fast_zero/app.py'
pre_test = 'task format'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "freezegun>=1.5.1",
    "pytest-asyncio>=0.25.3",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "taskipy>=1.13.0",
    "pytest-sugar>=1.0.0",
    "greenlet>=3.1.1",
    "ipdb>=0.13.13",
]

[tool.coverage.run]
concurrency = ["thread", "greenlet"]
